Change: Move App.js to container
