+++ src/container/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { inject, observer } from 'mobx-react';
-import LoginView from './container/LoginView'
-import ApplicationView from './container/ApplicationView'
+import LoginView from './LoginView'
+import ApplicationView from './ApplicationView'
 
 const App = ({store, event}) => (
   store.credential.isLoggedIn ? <ApplicationView user={store.credential.status} application={store.application} /> : <LoginView credential={store.credential} />
